# Implement section caching to speed up SIR body decoding.