Profiling suggested that quite a bit of time was spent looking up
sections in the executable (by name). This change caches references to
each section's bytes during initialisation (as we load the headers).

Rough benchmarking of the test suite:

Before:
```
            Mean        Std.Dev.    Min         Median      Max
real        4.810       0.285       4.463       4.637       5.230
user        22.695      0.245       22.242      22.679      23.145
sys         10.872      0.634       9.409       11.072      11.698
```

After:
```
            Mean        Std.Dev.    Min         Median      Max
real        3.874       0.156       3.689       3.869       4.094
user        11.645      0.165       11.279      11.707      11.825
sys         10.790      0.402       9.932       10.759      11.573
```