@ltratt 2020-12-09
That's a huge change in `user` time! Can you do me a favour and benchmark this with `--test-threads=1` before and after, as I think that'll give us a better picture of what's really going on (the `sys` time seems hugely high, but maybe that's a consequence of `mmap`?).