@vext01 2020-12-09
Before:
```
1: env RUST_TEST_THREADS=1 cargo test
            Mean        Std.Dev.    Min         Median      Max
real        14.841      0.059       14.751      14.856      14.931      
user        14.525      0.032       14.487      14.519      14.589      
sys         2.238       0.058       2.141       2.225       2.330       
```

After:
```
1: env RUST_TEST_THREADS=1 cargo test
            Mean        Std.Dev.    Min         Median      Max
real        7.104       0.023       7.068       7.104       7.159
user        7.161       0.053       7.080       7.166       7.248
sys         1.058       0.044       0.997       1.061       1.128
```

What are you thinking?