@ltratt 2020-12-09
This gives us a much better picture of the speed-up because we're no longer measuring accidental things like how many cores we have. 2x is not bad at all!