@vext01 2020-12-09
> 2x is not bad at all!

Better than a kick in the teeth!